py/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
mysql-connector-python==8.2.0
requests==2.31.0
beautifulsoup4==4.12.2
lxml==4.9.3
pandas==2.1.3
numpy==1.24.3
scipy==1.11.4
scikit-learn==1.3.2
matplotlib==3.8.2
seaborn==0.13.0
konlpy==0.6.0
gensim==4.3.2
jamo==0.4.1
torch==2.1.1
transformers==4.35.2
sentence-transformers==2.2.2
kobert-transformers==0.5.1
python-dotenv==1.0.0
pydantic==2.5.0
python-multipart==0.0.6
sqlalchemy==2.0.23
        